extract ast tests from esmodule-convert
