+++ packages/app/src/sandbox/eval/transpilers/babel/ast/convert-esmodule.test.ts
@@ -1,6 +1,5 @@
 import { convertEsModule as convert } from './convert-esmodule';
 import { generateCode, parseModule } from './utils';
-import { getSyntaxInfoFromAst } from './syntax-info';
 
 function convertEsModule(code: string) {
   const ast = parseModule(code);
@@ -8,8 +7,7 @@ function convertEsModule(code: string) {
   const result = {
     deps,
     code: generateCode(ast),
-  }
-  console.log(result);
+  };
   return result;
 }
 
@@ -434,7 +432,8 @@ describe('convert-esmodule', () => {
     expect(convertEsModule(code)).toMatchSnapshot();
   });
 
-  it.only('can convert import expressions', () => {
+  // TODO: We should probably also have asyncDeps in the result?
+  it('can convert import expressions', () => {
     const code = `
     import('test');
     `;
@@ -467,11 +466,11 @@ describe('convert-esmodule', () => {
 
     const result = convertEsModule(code);
     expect(result).toMatchSnapshot();
-    expect(result.indexOf('tfjs-core')).toBeLessThan(
-      result.indexOf('tfjs-data')
+    expect(result.code.indexOf('tfjs-core')).toBeLessThan(
+      result.code.indexOf('tfjs-data')
     );
-    expect(result.indexOf('tfjs-core')).toBeLessThan(
-      result.indexOf('tfjs-layers')
+    expect(result.code.indexOf('tfjs-core')).toBeLessThan(
+      result.code.indexOf('tfjs-layers')
     );
   });
 
@@ -483,8 +482,8 @@ describe('convert-esmodule', () => {
 
     const result = convertEsModule(code);
     expect(result).toMatchSnapshot();
-    expect(result.indexOf('tfjs-core')).toBeLessThan(
-      result.indexOf('tfjs-data')
+    expect(result.code.indexOf('tfjs-core')).toBeLessThan(
+      result.code.indexOf('tfjs-data')
     );
   });
 
@@ -498,11 +497,11 @@ describe('convert-esmodule', () => {
     `;
 
     const result = convertEsModule(code);
-    expect(result.indexOf('function a')).toBeGreaterThan(
-      result.indexOf('tfjs-data')
+    expect(result.code.indexOf('function a')).toBeGreaterThan(
+      result.code.indexOf('tfjs-data')
     );
-    expect(result.indexOf('tfjs-core')).toBeLessThan(
-      result.indexOf('tfjs-data')
+    expect(result.code.indexOf('tfjs-core')).toBeLessThan(
+      result.code.indexOf('tfjs-data')
     );
   });
 
@@ -517,8 +516,12 @@ describe('convert-esmodule', () => {
     `;
     const result = convertEsModule(code);
     expect(result).toMatchSnapshot();
-    expect(result.indexOf('function a')).toBeGreaterThan(result.indexOf('./a'));
-    expect(result.indexOf('function a')).toBeGreaterThan(result.indexOf('./b'));
+    expect(result.code.indexOf('function a')).toBeGreaterThan(
+      result.code.indexOf('./a')
+    );
+    expect(result.code.indexOf('function a')).toBeGreaterThan(
+      result.code.indexOf('./b')
+    );
   });
 
   it('keeps star exports after default export order', () => {
@@ -530,8 +533,8 @@ describe('convert-esmodule', () => {
     `;
     const result = convertEsModule(code);
     expect(result).toMatchSnapshot();
-    expect(result.indexOf('./containers')).toBeGreaterThan(
-      result.indexOf('./withSearch')
+    expect(result.code.indexOf('./containers')).toBeGreaterThan(
+      result.code.indexOf('./withSearch')
     );
   });
 
@@ -590,45 +593,5 @@ export function test3() {
 
     const result = convertEsModule(code);
     expect(result).toMatchSnapshot();
-  });
-
-  describe('syntax info', () => {
-    it('can detect jsx', () => {
-      const code = `const a = <div>Hello</div>`;
-      const ast = parseModule(code);
-      convert(ast);
-      const syntaxInfo = getSyntaxInfoFromAst(ast);
-      expect(syntaxInfo.jsx).toBeTruthy();
-    });
-
-    it('can detect non jsx', () => {
-      const code = `const a = /<div>Hello<\\/div>/`;
-      const ast = parseModule(code);
-      convert(ast);
-      const syntaxInfo = getSyntaxInfoFromAst(ast);
-      expect(syntaxInfo.jsx).toBeFalsy();
-    });
-  });
-
-  describe('printer issues', () => {
-    it('can convert + +', () => {
-      const code = `
-      c = (10.0, + +(15))
-      `;
-
-      expect(convertEsModule(code)).toMatchSnapshot();
-    });
-
-    it('can convert -(--i)', () => {
-      const code = `a = -(--i)`;
-      expect(convertEsModule(code)).toBe('"use strict";\na = - --i;\n');
-    });
-
-    it('can convert unicode line breaks', () => {
-      const code = `const a = "[\\u2028]";`;
-      expect(convertEsModule(code)).toBe(
-        '"use strict";\nconst a = "[\\u2028]";\n'
-      );
-    });
   });
 });
